refactor skill list
