+++ packages/playground/src/domains/agent-builder/components/agent-edit/agent-profile/skills.tsx
@@ -1,13 +1,84 @@
 import type { StoredSkillResponse } from '@mastra/client-js';
-import { SkillsDetail } from '../details/skills-detail';
+import { Checkbox, Txt } from '@mastra/playground-ui';
+import { useFormContext } from 'react-hook-form';
+import type { AgentBuilderEditFormValues } from '../../../schemas';
 
-export interface SkillsProps {
-  availableSkills: StoredSkillResponse[];
+interface SkillsProps {
   editable?: boolean;
-  /** Disables interaction (e.g. while a stream is running). */
-  disabled?: boolean;
+  availableSkills?: StoredSkillResponse[];
 }
 
-export const Skills = ({ availableSkills, editable = true, disabled = false }: SkillsProps) => {
-  return <SkillsDetail editable={editable && !disabled} availableSkills={availableSkills} />;
+export const Skills = ({ editable = true, availableSkills = [] }: SkillsProps) => {
+  const { setValue, getValues, watch } = useFormContext<AgentBuilderEditFormValues>();
+  const selected = watch('skills') ?? {};
+
+  const toggle = (id: string, next: boolean) => {
+    const current = getValues('skills') ?? {};
+    setValue('skills', { ...current, [id]: next }, { shouldDirty: true });
+  };
+
+  return (
+    <div className="h-full min-h-0 overflow-y-auto py-2">
+      {availableSkills.length === 0 ? (
+        <SkillEmptyState />
+      ) : (
+        <ul className="flex flex-col">
+          {availableSkills.map(skill => {
+            const isChecked = Boolean(selected[skill.id]);
+
+            return (
+              <li key={skill.id}>
+                <SkillItem skill={skill} editable={editable} onToggle={toggle} isChecked={isChecked} />
+              </li>
+            );
+          })}
+        </ul>
+      )}
+    </div>
+  );
+};
+
+interface SkillItemProps {
+  skill: StoredSkillResponse;
+  editable: boolean;
+  onToggle: (id: string, next: boolean) => void;
+  isChecked: boolean;
+}
+
+const SkillItem = ({ skill, editable, onToggle, isChecked }: SkillItemProps) => {
+  return (
+    <label
+      className="flex cursor-pointer items-start gap-3 px-6 py-4 transition-colors hover:bg-surface2"
+      aria-disabled={!editable}
+    >
+      <div className="mt-0.5">
+        <Checkbox
+          variant="neutral"
+          checked={isChecked}
+          onCheckedChange={next => onToggle(skill.id, next === true)}
+          disabled={!editable}
+        />
+      </div>
+      <div className="flex min-w-0 flex-col">
+        <Txt variant="ui-sm" className="font-medium text-neutral6">
+          {skill.name}
+        </Txt>
+        {skill.description && (
+          <Txt variant="ui-xs" className="mt-0.5 truncate text-neutral3" title={skill.description}>
+            {skill.description}
+          </Txt>
+        )}
+      </div>
+    </label>
+  );
+};
+
+const SkillEmptyState = () => {
+  return (
+    <div className="flex items-center justify-center py-12">
+      <Txt variant="ui-sm" className="text-neutral3">
+        No skills available in this project.
+      </Txt>
+    </div>
+  );
 };
